Because all builtin pytypes are singletons created first at runtime, both them and their properties are not reset every time you create a new ModuleEnv(), like in the unittest testcases. Because of this, types passed to any of these will persist between new instances and parsing of ModuleEnv().
This behavior is seen between 2 unittests in the test_list module where
one test accesses a list by indexing and another by slicing. As a
result, the environment of the ListGetItemMethod saves both an int and
slice as the "key" variable.

Running the unittests now will not show this error because the
adjusted_call() method was what was overrided when subclassing
GetItemMethod. The return types of adjusted_call() were only based off
the arguments passed to the function during a call and ignores any
previous values passed to it from another ModuleEnv. This behavior can
be seen by instead overriding the returns() method where you are forced
to use the enviornment to get the argument types.

For the 2 tests, the indexing test happens first, so key will just be an
int, but the second test that comes after uses a slice as the key. So
overriding returns() and calling self.env().lookup("key") returns the
set of an INT_TYPE and SLICE_TYPE.

To fix this, the types must either somehow be attatched to the ModuleEnv
so the types and their properties get reset every time you use a new
ModuleEnv. Alternatively, we could create a new class that contains
this ModulEnv and the builtin types that we can create on multiple
instances during the python interpretter lifetime. Either way, A LOT of
code will need to be changed such that none of the builtin types can be
imported globally and need to be accesed from this parent class because
they should no longer be singletons if we want to run multiple instances
of ModuleEnv in unittests.